apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
openai>=1.0.0
PyMuPDF>=1.23.0
pymupdf4llm>=0.0.10
playwright>=1.40.0
markdown2>=2.4.0

# OCR dependencies (choose one)
# Option 1: PaddleOCR (recommended)
paddleocr>=2.7.0
paddlepaddle>=2.6.0

# Option 2: Tesseract (alternative)
# pytesseract>=0.3.10
# Pillow>=10.0.0